Updated coding blog: jet&snake, readme
