+++ pages/works/coding/readme/index.js
@@ -6,6 +6,9 @@ import cssClasses from "./index.module.css";
 
 import backIcon from "/img/others/back.png";
 
+import Latex from 'react-latex-next';
+import 'katex/dist/katex.min.css';
+
 export default function Home() {
   return (
     <div className={cssClasses.canvas}>
@@ -26,36 +29,22 @@ export default function Home() {
           </Link>
 
           <p>
-            This is my personal website for me to put my works and drawings
-            online for people to have a look. I also made a blog page to
-            motivate myself to write some study blogs on maths, physics and
-            programming. Wish me not being too lazy or busy so I can keep
-            friends and peers updated.
+          This is a very informal personal website for me to share my works and drawings for people to have a look. There is also a blog page for me to record my study progress on maths, physics and programming.
           </p>
           <p>
-            This website is running on the Next.js framework. The source code
-            was put on GitHub then fetched and deployed on Vercel&rsquo;s
-            platform, which is the recommended deployment method for Next.js
-            project.
+          This website is running on the Next.js framework. The source code repo is public on GitHub to be fetched and deployed by Vercel.
           </p>
           <p>
-            The name of the page - Room 214N, was inspired by the room number I
-            stayed at Connaught Hall in my first 2 years of undergraduate. In
-            fact I was there in my third year as well but in a different room.
-            There is another reference in the choice. Try to see if you can
-            figure it out.
+          The name of this page - Room 214N, was inspired by my Connaught Hall room number during the first 2 years of my undergraduate study. In fact I was there in my third year as well but a different room. Try to see if you can figure out why there is an “N” at the end.
           </p>
           <p>
-            I have been trash talking so far cuz this Readme page is actually a
-            blog template for me to make updates. I can just duplicate this
-            directory, make a link on the parent page, and copy my written
-            content in.
+          I have been trash talking so far because this Readme page is actually a blog template for me to make updates. I can just duplicate this directory, make a link on the parent index page, and copy my written contents in.
           </p>
           <p>
-            Someday I will add an interface in Chinese as well. I want it to
-            happen but I don&rsquo;t want to do it myself :( There will be so
-            much to rewrite and the amount of work will double when I am
-            updating. Sorry mum ask dad or Google to translate...
+          Some day I will add an interface in Chinese as well. Some day...
+          </p>
+          <p>
+          Nov 2023 update: I can now type <Latex>$\LaTeX$</Latex> in my blog. I am happy :)
           </p>
         </div>
       </div>
